move dict to new constants file
